migrate messaging to adapter
